fix: more -Wextra fixes

+++ bpf/egress.c
@@ -34,7 +34,7 @@ static inline int mangle_data(struct __sk_buff* skb, __u16 offset, __be32* csum_
     bpf_gt0_hack2(padding_len);
     padding_len = min(padding_len, MAX_PADDING_LEN);
 
-    for (int i = 0; i < padding_len / 4 + !!(padding_len % 4); i++)
+    for (size_t i = 0; i < padding_len / 4 + !!(padding_len % 4); i++)
       ((__u32*)buf)[i] = bpf_get_prandom_u32();
     // HACK: prevent usage of __builtin_memset against variable size
     switch (padding_len % 4) {
